issue #9 enhancement
Create a new file in the utils folder called facts.py and write a function called random_facts that calls up random facts about cats

In the cat class in cat.py and in the owner class in human.py, create a function called give_fact that pulls a random fact about cats.

Write documentation about the function

Write a test for the function.

 Changes to be committed:
	modified:   ../api/cat.py
	modified:   ../api/human.py
	modified:   test_api_Cat.py
	modified:   test_api_Owner.py
	modified:   ../utils/__init__.py
	new file:   ../utils/facts.py